+++ src/main/java/br/com/sample/api/voting/manager/java/service/IVotingSessionService.java
@@ -1,5 +1,5 @@
 package br.com.sample.api.voting.manager.java.service;
 
 public interface IVotingSessionService {
-    void openVotingSession(Long id, Integer duration);
+    void openVotingSession(Long id, Long duration);
 }
